transcription-app/requirements.txt
# =============================================================================
# Application de Transcription Audio Offline
# Configuration: Intel i7 / 16 Go RAM / CPU-only
# =============================================================================

# --- Transcription (faster-whisper + CTranslate2) ---
faster-whisper>=1.0.3
ctranslate2>=4.4.0

# --- Diarization ---
pyannote.audio>=3.1.1
# Alternative rapide (décommenter si besoin)
# speechbrain>=1.0.0

# --- Audio processing ---
pydub>=0.25.1
sounddevice>=0.4.7
soundfile>=0.12.1
numpy>=1.26.0

# --- UI (PySide6) ---
PySide6>=6.6.0

# --- Utilitaires ---
pyyaml>=6.0.1
huggingface-hub>=0.20.0
tqdm>=4.66.0

# --- Dev & Tests ---
pytest>=8.0.0
pytest-qt>=4.4.0

# --- Packaging (installer séparément selon besoin) ---
# pyinstaller>=6.3.0
# nuitka>=2.0.0
